Added fast-commit support for dax
With this commit, now fast-commit works for dax devices in a
synchronous mode. If fast-commits are enabled, and the device is
mounted as dax, fsync() becomes a no-op and the updates are tracked
when they happen instead of waiting till fsync. Also added ops
to track coow operations.

Performance: 14m ops on filebench
TODOs:
1) Replay dummy inode with real inode.
2) Implement recovery code.